:recycle: Refactor: cx 분리
